+++ app/controllers/research_controller.rb
@@ -1,14 +1,13 @@
 class ResearchController < ApplicationController
   before_action :authenticate_user!, :only => [:research_karma, :research_surveys]
-  before_action :set_active_top_nav_link_to_research
-  before_action :authenticate_research, only: [:research_surveys]
-  
-  def research_surveys
-  end
+  before_action :set_active_top_nav_link_to_surveys
 
+  before_action :fetch_notifications
 
-  def research_today
-    @coming_soon = true
+  layout "community"
+
+  def fetch_notifications
+    @posts = Post.notifications.viewable.all
   end