+++ app/controllers/surveys_controller.rb
@@ -1,6 +1,8 @@
 class SurveysController < ApplicationController
   before_filter :authenticate_user!
-  before_action :authenticate_research
+  before_action :set_active_top_nav_link_to_surveys
+
+  layout "community"
 
   def start_survey
     @question_flow = QuestionFlow.find(params[:question_flow_id])