+++ app/models/surveys/answer_options_answer_template.rb
@@ -1,4 +1,0 @@
-class AnswerOptionsAnswerTemplate < ActiveRecord::Base
-  belongs_to :answer_template
-  belongs_to :answer_option
-end