+++ app/models/surveys/answer_options_template.rb
@@ -1,0 +1,4 @@
+class AnswerOptionsTemplate < ActiveRecord::Base
+  belongs_to :answer_template
+  belongs_to :answer_option
+end