+++ config/routes.rb
@@ -1,5 +1,12 @@
 Rails.application.routes.draw do
 
+  # This line mounts Forem's routes at /forums by default.
+  # This means, any requests to the /forums URL of your application will go to Forem::ForumsController#index.
+  # If you would like to change where this extension is mounted, simply change the :at option to something different.
+  #
+  # We ask that you don't use the :as option here, as Forem relies on it being the default of "forem"
+  mount Forem::Engine, :at => '/social/discussion'
+
   # The priority is based upon order of creation: first created -> highest priority.
   # See how all your routes lay out with "rake routes".
 
@@ -9,6 +16,9 @@ Rails.application.routes.draw do
   get 'external_link_warning' => 'static#external_link_warning'
   get 'theme' => 'static#theme'
   get 'version' => 'static#version'
+  #Content Pages
+  get 'content/:page' => 'static#content'
+  get 'content/' => 'static#content'
 
   # MyApnea Specific
   get 'learn' => 'static#learn'
@@ -105,6 +115,7 @@ Rails.application.routes.draw do
   # We ask that you don't use the :as option here, as Forem relies on it being the default of "forem"
   mount Forem::Engine, :at => '/social/discussion'
 
+
 # # Authentication
 #   devise_for :user, skip: [:sessions, :passwords, :confirmations, :registrations]
 #   as :user do
@@ -187,7 +198,4 @@ Rails.application.routes.draw do
   #     # (app/controllers/admin/products_controller.rb)
   #     resources :products
   #   end
-
-
-
 end