+++ db/migrate/20140325050821_create_answer_options_templates.rb
@@ -1,0 +1,10 @@
+class CreateAnswerOptionsTemplates < ActiveRecord::Migration
+  def change
+    create_table :answer_options_templates do |t|
+      t.references :answer_template
+      t.references :answer_option
+
+      t.timestamps
+    end
+  end
+end